Properly support date and time type incremental bookmarks
Motivation
----------

https://stitchdata.atlassian.net/browse/SRCE-722

We added support for date and time type incremental bookmarks for the
full table syncs in dc4a0e0 and following. Unfortunately, we couldn't
write these types as bookmarks because they were not formatted as strings.

This change stores them as strings when they're generated and then
everywhere else can just use them as is.